+++ src/afw/os/win/afw_os.c
@@ -25,9 +25,25 @@
 #include <time.h>
 #include <signal.h>
 
-static const afw_utf8_t
-impl_s_no_memory_for_backtrace =
-    AFW_UTF8_LITERAL("No memory for backtrace");
+static const afw_memory_t impl_no_memory_for_backtrace = {
+    (const afw_octet_t *)"No memory for backtrace",
+    sizeof("No memory for backtrace") - 1
+};
+
+static const afw_memory_t *
+impl_memory_unhandled(
+    const afw_octet_t *ptr, afw_size_t size, afw_xctx_t *xctx)
+{
+    afw_memory_t *m;
+
+    m = afw_pool_malloc_unhandled(xctx->p, sizeof(afw_memory_t), xctx);
+    if (!m) {
+        return &impl_no_memory_for_backtrace;
+    }
+    m->ptr = ptr;
+    m->size = size;
+    return m;
+}
 
 static const afw_utf8_t
 impl_dso_suffix =
@@ -197,15 +213,15 @@ afw_os_get_dso_suffix()
 
 
 /*  Provide a backtrace if possible. */
-AFW_DEFINE(const afw_utf8_t *)
+AFW_DEFINE(const afw_memory_t *)
 afw_os_backtrace(
     afw_error_code_t code,
     int max_backtrace,
     afw_xctx_t *xctx)
 {
 
-    afw_utf8_t *result;
-    afw_octet_t *s;
+    char *s;
+    char *start;
     afw_size_t len;
     int wlen;
 
@@ -244,10 +260,8 @@ afw_os_backtrace(
      */
     len = (max + 10) * 100;
     s = afw_pool_malloc_unhandled(xctx->p, len, xctx);
-    if (!s) return &impl_s_no_memory_for_backtrace;
-    result = afw_pool_malloc_unhandled(xctx->p, sizeof(afw_utf8_t), xctx);
-    if (!result) return &impl_s_no_memory_for_backtrace;
-    result->s = s;
+    if (!s) return &impl_no_memory_for_backtrace;
+    start = s;
 
     /* Capture backtrace stack. */
     frames = CaptureStackBackTrace(0, max, stack, NULL);
@@ -282,8 +296,8 @@ afw_os_backtrace(
         }
     }
 
-    result->len = s - result->s;
-    return result;
+    return impl_memory_unhandled(
+        (const afw_octet_t *)start, (afw_size_t)(s - start), xctx);
 }